Appending Writer implementation
This breaks the read/write functionality, and has been implemented using a hack
to avoid reading before writing. It is therefore not a general Writer
implementation, but demonstrates what is required to support write appending.
Although performance has not been measured, the simple writer-test suite
certainly seems to perform faster.