Add macro to control whether a single large file datatype is used
Current implementation of fileview datatype is to create one datatype that is a
contiguous datatype formed from a more elementary datatype, large enough that it
can span the entire file. It's not clear why this is necessary, but without
doing this some test cases fail, in particular when appending to a file.