Refactor DataDistribution based on "block generators"
This generalizes DataDistribution to support a much wider variety of
distributions. In particular, there is support for distribution by a sequence of
enumerated index values (or blocks). Block generators are functions that take a
state and return a new state and (optionally) a block. A capability was added to
create blocks using an iterator of the data distribution for a given number of
elements (indices), allowing the original blocks to be split.