GridDistribution for QuasiEquipartitionGenerator
Since a GridDistribution is a function taking an optional axis length, and
QuasiEquipartitionGenerator must have a fixed axis length, it's uncertain what
to do when the GridDistribution function is called with an empty axis value.
This implementation chooses to return a cyclic data distribution with a block
length of one in that case. The choice of cyclic distribution was made since it
will result in approximately the same number of axis values per rank as a
quasiequipartitioned distribution once the axis length is fixed, and sequential
access to the indeterminate length axis will be fairly evenly distributed across
ranks regardless of the final length.